+++ src/Middleware/DispatcherInterface.php
@@ -16,17 +16,6 @@ use Psr\Http\Message\ServerRequestInterface;
 interface DispatcherInterface extends \Interop\Http\ServerMiddleware\MiddlewareInterface
 {
     /**
-     * Processes the specified middlewares from stack.
-     * NOTE: To be removed in v1.0.0. Use MiddlewareInterface::process instead.
-     *
-     * @param  \Psr\Http\Message\RequestInterface  $request
-     * @param  \Psr\Http\Message\ResponseInterface $response
-     * @param  array                               $stack
-     * @return \Psr\Http\Message\ResponseInterface
-     */
-    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, array $stack = array());
-
-    /**
      * Adds a new middleware or a list of middlewares in the stack.
      *
      * @param  callable|object|string|array $middleware